added swapping algos mechanism

+++ src/main/java/model/Algos/BestFitAlgo.java
@@ -5,7 +5,7 @@ import model.Process;
 import java.util.Optional;
 
 public class BestFitAlgo extends FitAlgo {
-    public BestFitAlgo(int totalmem) {
+    public BestFitAlgo(long totalmem) {
         super(totalmem);
         name = "Best Fit";
     }
