+++ src/main/java/model/Algos/FirstFitAlgo.java
@@ -10,7 +10,7 @@ public class FirstFitAlgo extends FitAlgo {
 
 
     //divides total memory to make into kilobytes.
-    public FirstFitAlgo(int totalmem){
+    public FirstFitAlgo(long totalmem){
         super(totalmem);
         name = "First Fit";
     }
