+++ src/main/java/model/Algos/WorstFitAlgo.java
@@ -5,7 +5,7 @@ import model.Process;
 import java.util.Optional;
 
 public class WorstFitAlgo extends FitAlgo {
-    public WorstFitAlgo(int totalmem) {
+    public WorstFitAlgo(long totalmem) {
         super(totalmem);
         name = "Worst Fit";
     }
